Visualization/requirements.txt
PyQt6
Pillow
opencv-python
torch
torchvision
utils
